- Use strlcpy, strlcat instead of strnzcpy and strnzcat. - Replace strnzncpy and strnzncat with strlncpy and strlncat which are expressed in terms of strlcat and strlcat. - Add strlcpy and strlcat to compat/. - Move most of compat #ifdef's to compat/. - Remove most of unneeded casts to (size_t) and (int).
git-svn-id: file:///home/fjoe/old-repos/svn/husky/trunk/sog@4360 1738b366-ca78-dd11-ac90-000e8e03bdb1

+++ src/mod_bootdb/db_area.c
@@ -1,5 +1,5 @@
 /*
- * $Id: db_area.c,v 1.148 2003-05-14 17:41:59 fjoe Exp $
+ * $Id: db_area.c,v 1.149 2003-09-29 23:11:25 fjoe Exp $
  */
 
 /***************************************************************************
@@ -789,7 +789,7 @@ DBLOAD_FUN(load_specials)
 				break;
 			}
 
-			while (ssubst - spec_substs < (int) SPEC_SUBSTS_SZ &&
+			while ((size_t) (ssubst - spec_substs) < SPEC_SUBSTS_SZ &&
 			       !str_cmp(spec, ssubst->spec)) {
 				trig = trig_new(
 				    &pMobIndex->mp_trigs, ssubst->trig);
